WEB-4208: forgot intro steps component
